Add sound files, change fps
Changed the fps so that the button presses are more manageable. This
means no more flashing inventory screen.
Changed map generation to run a while loop rather than a range loop.
There where a lot of discarded room placements.
Put limits on the random room placement so that if they are near the
edges of the map a new position is chosen.
Turned on room ID while in debug.
Some attempt at monster balance, really it is just random numbers.
reformat item and monster placement code to be more readable.