added test environment file
Test.py is a one room test environment to see how objects and creatures
work

Test_color_cycle.py is a color cycling of the background based on
rendering steps. could possibly use in showing health or indicating
health.

Test_speed_system.py is a working environment where a monster can be
faster or slower than the player. currently speed is ramped up pretty
extreme to be able to see it work.